Rename 'nodes' module 'prospective'
