+++ src/flowrep/api/schemas.py
@@ -21,18 +21,18 @@ from flowrep.edge_models import OutputEdges as OutputEdges
 from flowrep.edge_models import OutputTarget as OutputTarget
 from flowrep.edge_models import SourceHandle as SourceHandle
 from flowrep.edge_models import TargetHandle as TargetHandle
-from flowrep.nodes.atomic_recipe import AtomicRecipe as AtomicRecipe
-from flowrep.nodes.atomic_recipe import UnpackMode as UnpackMode
-from flowrep.nodes.for_recipe import ForEachRecipe as ForEachRecipe
-from flowrep.nodes.helper_models import ConditionalCase as ConditionalCase
-from flowrep.nodes.helper_models import ExceptionCase as ExceptionCase
-from flowrep.nodes.helper_models import LabeledRecipe as LabeledRecipe
-from flowrep.nodes.if_recipe import IfRecipe as IfRecipe
-from flowrep.nodes.try_recipe import TryRecipe as TryRecipe
-from flowrep.nodes.union_types import RecipeDiscrimination as RecipeDiscrimination
-from flowrep.nodes.union_types import Recipes as Recipes
-from flowrep.nodes.while_recipe import WhileRecipe as WhileRecipe
-from flowrep.nodes.workflow_recipe import WorkflowRecipe as WorkflowRecipe
+from flowrep.prospective.atomic_recipe import AtomicRecipe as AtomicRecipe
+from flowrep.prospective.atomic_recipe import UnpackMode as UnpackMode
+from flowrep.prospective.for_recipe import ForEachRecipe as ForEachRecipe
+from flowrep.prospective.helper_models import ConditionalCase as ConditionalCase
+from flowrep.prospective.helper_models import ExceptionCase as ExceptionCase
+from flowrep.prospective.helper_models import LabeledRecipe as LabeledRecipe
+from flowrep.prospective.if_recipe import IfRecipe as IfRecipe
+from flowrep.prospective.try_recipe import TryRecipe as TryRecipe
+from flowrep.prospective.union_types import RecipeDiscrimination as RecipeDiscrimination
+from flowrep.prospective.union_types import Recipes as Recipes
+from flowrep.prospective.while_recipe import WhileRecipe as WhileRecipe
+from flowrep.prospective.workflow_recipe import WorkflowRecipe as WorkflowRecipe
 from flowrep.retrospective import NOT_DATA as NOT_DATA
 from flowrep.retrospective import AtomicData as AtomicData
 from flowrep.retrospective import CompositeData as CompositeData
