+++ tests/integration/parsers/test_parsing_if_nodes.py
@@ -1,7 +1,7 @@
 import unittest
 
-from flowrep.nodes import workflow_recipe
 from flowrep.parsers import workflow_parser
+from flowrep.prospective import workflow_recipe
 
 from flowrep_static import library
