+++ tests/integration/parsers/test_parsing_try_nodes.py
@@ -2,8 +2,8 @@ import unittest
 
 from pyiron_snippets import versions
 
-from flowrep.nodes import workflow_recipe
 from flowrep.parsers import workflow_parser
+from flowrep.prospective import workflow_recipe
 
 from flowrep_static import library
