+++ tests/unit/nodes/test_if_recipe.py
@@ -3,7 +3,7 @@ import unittest
 import pydantic
 
 from flowrep import base_models, edge_models, subgraph_validation
-from flowrep.nodes import (
+from flowrep.prospective import (
     atomic_recipe,
     helper_models,
     if_recipe,
