+++ tests/unit/nodes/test_try_recipe.py
@@ -4,7 +4,7 @@ import pydantic
 from pyiron_snippets import versions
 
 from flowrep import base_models, edge_models, subgraph_validation
-from flowrep.nodes import (
+from flowrep.prospective import (
     atomic_recipe,
     helper_models,
     try_recipe,
