+++ tests/unit/nodes/test_union_types.py
@@ -6,7 +6,7 @@ import pydantic
 from pyiron_snippets import versions
 
 from flowrep import base_models
-from flowrep.nodes import (
+from flowrep.prospective import (
     atomic_recipe,
     for_recipe,
     if_recipe,
