+++ tests/unit/nodes/test_workflow_recipe.py
@@ -3,7 +3,7 @@ import unittest
 import pydantic
 
 from flowrep import base_models, edge_models, subgraph_validation
-from flowrep.nodes import atomic_recipe, workflow_recipe
+from flowrep.prospective import atomic_recipe, workflow_recipe
 
 from flowrep_static import makers
