+++ tests/unit/parsers/test_try_parser.py
@@ -3,14 +3,14 @@ import unittest
 from pyiron_snippets import versions
 
 from flowrep import edge_models
-from flowrep.nodes import (
+from flowrep.parsers import try_parser, workflow_parser
+from flowrep.prospective import (
     for_recipe,
     if_recipe,
     try_recipe,
     while_recipe,
     workflow_recipe,
 )
-from flowrep.parsers import try_parser, workflow_parser
 
 from flowrep_static import library
