+++ tests/unit/parsers/test_while_parser.py
@@ -1,14 +1,14 @@
 import unittest
 
 from flowrep import edge_models
-from flowrep.nodes import (
+from flowrep.parsers import while_parser, workflow_parser
+from flowrep.prospective import (
     for_recipe,
     if_recipe,
     try_recipe,
     while_recipe,
     workflow_recipe,
 )
-from flowrep.parsers import while_parser, workflow_parser
 
 from flowrep_static import library
